requirements.txt
google-cloud-vision
Flask
pandas